% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankingPlots.R
\name{barPlot}
\alias{barPlot}
\title{Bar plot}
\usage{
barPlot(
  experiment,
  assay = 1,
  annotate = NULL,
  scaling = "auto",
  log = exp(1),
  transpose = FALSE
)
}
\arguments{
\item{experiment}{SummarizedExperiment object with at least
one (matrix) assay.}

\item{assay}{Name or index of the assay to plot. Defaults to
the first assay in [assayNames].}

\item{scaling}{Should autoscaling be performed on the assay?
Defaults to `TRUE`.}

\item{log}{Should the data be log transformed?
Defaults to `TRUE`}
}
\description{
Bar plot
}
