% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationPlots.R
\name{heatmapPlot}
\alias{heatmapPlot}
\title{heatmapPlot}
\usage{
heatmapPlot(
  experiment,
  assay = 1,
  scaling = "auto",
  log = exp(1),
  rows = NULL,
  columns = NULL,
  ...
)
}
\arguments{
\item{experiment}{SummarizedExperiment object with at least
one (matrix) assay.}

\item{assay}{Name or index of the assay to plot. Defaults to
the first assay in [assayNames].}
}
\description{
heatmapPlot
}
