% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepareData}
\alias{prepareData}
\title{Prepare data for plotting}
\usage{
prepareData(experiment, assay = 1, autoScale = TRUE, logTransform = TRUE)
}
\arguments{
\item{experiment}{SummarizedExperiment object with at least
one (matrix) assay.}

\item{assay}{Name or index of the assay to plot. Defaults to
the first assay in [assayNames].}

\item{autoScale}{Should autoscaling be performed on the assay?
Defaults to `TRUE`.}

\item{logTransform}{Should the data be log transformed?
Defaults to `TRUE`}
}
\description{
Prepare data for plotting
}
